server/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate
Flask-Cors
Flask-Bcrypt
Flask-JWT-Extended==4.6.0
SQLAlchemy-Serializer
psycopg2-binary==2.9.9
Werkzeug==3.0.1
python-dotenv==1.0.0
marshmallow==3.20.1